# docs/10_api_openapi.yaml
openapi: 3.0.3
info:
  title: Freight Forwarding and Finance Platform API
  version: 2.0.0
  description: >-
    Organization- and branch-scoped API for freight operations, informational
    service charges, reusable financial documents, and double-entry accounting.
servers:
  - url: /api/v1
    description: Current API
security:
  - bearerAuth: []
...
paths:
...
  /financial-documents:
    get:
      tags: [Finance]
      operationId: listFinancialDocuments
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/BranchFilter'}, {name: document_type, in: query, schema: {type: string}}, {name: status, in: query, schema: {type: string}}, {name: party_id, in: query, schema: {type: integer}}]
      responses: {'200': {description: Financial documents, content: {application/json: {schema: {$ref: '#/components/schemas/PagedFinancialDocumentResponse'}}}}}
    post:
      tags: [Finance]
      operationId: createFinancialDocument
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/FinancialDocumentCreate'}}}
      responses: {'201': {description: Draft document created}}
  /financial-documents/{documentId}/post:
    post:
      tags: [Finance]
      operationId: postFinancialDocument
      parameters: [{$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '200': {description: Posted and journal generated}
        '409': {description: Period closed, already posted, or duplicate command}
  /financial-documents/{documentId}/allocate:
    post:
      tags: [Finance]
      operationId: allocatePayment
      parameters: [{$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/AllocationCreate'}}}
      responses: {'201': {description: Allocation created}}
  /financial-documents/{documentId}/reverse:
    post:
      tags: [Finance]
      operationId: reverseFinancialDocument
      parameters: [{$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ReversalRequest'}}}
      responses: {'201': {description: Reversal created}}
  /journal-entries:
    get:
      tags: [Accounting]
      operationId: listJournalEntries
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/BranchFilter'}, {name: status, in: query, schema: {type: string}}]
      responses: {'200': {description: Journal entries}}
    post:
      tags: [Accounting]
      operationId: createManualJournal
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/JournalEntryCreate'}}}
      responses: {'201': {description: Draft journal created}}
  /journal-entries/{journalId}/post:
    post:
      tags: [Accounting]
      operationId: postManualJournal
      parameters: [{$ref: '#/components/parameters/JournalId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: Balanced journal posted}}
  /accounting-periods/{periodId}/close:
    post:
      tags: [Accounting]
      operationId: closeAccountingPeriod
      parameters: [{$ref: '#/components/parameters/PeriodId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ReasonRequest'}}}
      responses: {'200': {description: Period closed}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    DocumentId: {name: documentId, in: path, required: true, schema: {type: integer, format: int64}}
    JournalId: {name: journalId, in: path, required: true, schema: {type: integer, format: int64}}
    PeriodId: {name: periodId, in: path, required: true, schema: {type: integer, format: int64}}
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
    BranchFilter: {name: branch_id, in: query, schema: {type: integer, format: int64}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
...
  schemas:
...
    PageMeta:
      type: object
      properties: {page: {type: integer}, page_size: {type: integer}, total: {type: integer}}
...
    FinancialDocumentCreate:
      type: object
      required: [document_type, document_date, currency_code, lines]
      properties: {document_type: {type: string}, document_date: {type: string, format: date}, party_id: {type: integer}, service_order_id: {type: integer}, branch_id: {type: integer}, currency_code: {type: string, minLength: 3, maxLength: 3}, description: {type: string}, reference_number: {type: string}, due_date: {type: string, format: date}, payment_method_code: {type: string}, financial_account_id: {type: integer}, lines: {type: array, minItems: 1, items: {type: object, required: [description, line_amount], properties: {description: {type: string}, fee_type_id: {type: integer}, service_order_id: {type: integer}, service_order_container_id: {type: integer}, account_id: {type: integer}, quantity: {type: number}, unit_price: {type: number}, line_amount: {type: number}}}}}
    AllocationCreate:
      type: object
      required: [target_document_id, allocated_amount, allocated_currency_code]
      properties: {target_document_id: {type: integer}, allocated_amount: {type: number, exclusiveMinimum: 0}, allocated_currency_code: {type: string, minLength: 3, maxLength: 3}, exchange_rate: {type: number, exclusiveMinimum: 0}}
    ReversalRequest:
      type: object
      required: [reason]
      properties: {reason: {type: string, minLength: 1}}
    JournalEntryCreate:
      type: object
      required: [entry_type, entry_date, posting_date, accounting_period_id, lines]
      properties: {entry_type: {type: string}, entry_date: {type: string, format: date}, posting_date: {type: string, format: date}, accounting_period_id: {type: integer}, description: {type: string}, branch_id: {type: integer}, lines: {type: array, minItems: 2, items: {type: object, required: [account_id, currency_code], properties: {account_id: {type: integer}, branch_id: {type: integer}, debit_amount: {type: number}, credit_amount: {type: number}, currency_code: {type: string, minLength: 3, maxLength: 3}, description: {type: string}}}}}
    ReasonRequest:
      type: object
      required: [reason]
      properties: {reason: {type: string}}
...
    FinancialDocument: {type: object, properties: {id: {type: integer}, document_no: {type: string}, document_type: {type: string}, status: {type: string}, total_amount: {type: number}}}
    PagedFinancialDocumentResponse: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/FinancialDocument'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}